# PERF: Series.any